Fix errors when try delete records that has two or more files fields in the same Model, It was not deleting the  second file and his folder

+++ Model/Behavior/UploadBehavior.php
@@ -389,22 +389,45 @@ class UploadBehavior extends ModelBehavior {
 		}
 
 		$folders = $this->__foldersToRemove[$model->alias];
-		$folder = $folders[0];
-	        $dir = $path . $folder;
-	        $it = new RecursiveDirectoryIterator($dir);
-	        $files = new RecursiveIteratorIterator($it, RecursiveIteratorIterator::CHILD_FIRST);
-	        foreach ($files as $file) {
-	            if ($file->getFilename() === '.' || $file->getFilename() === '..') {
-	                continue;
-	            }
-	
-	            if ($file->isDir()) {
-	                $this->rmdir($file->getRealPath());
-	            } else {
-	                $this->unlink($file->getRealPath());
-	            }
-	        }
-	        $this->rmdir($dir);
+        $folder = $folders[0];
+
+
+        $dir = $path . $folder;
+        $it = new RecursiveDirectoryIterator($dir);
+        $files = new RecursiveIteratorIterator($it, RecursiveIteratorIterator::CHILD_FIRST);
+        foreach ($files as $file) {
+            if ($file->getFilename() === '.' || $file->getFilename() === '..') {
+                continue;
+            }
+
+            if ($file->isDir()) {
+                $this->rmdir($file->getRealPath());
+            } else {
+                $this->unlink($file->getRealPath());
+            }
+        }
+        $this->rmdir($dir);
+		/*foreach ($folders as $folder) {
+			if (strlen((string)$folder) === 0) {
+				continue;
+			}
+
+			$dir = $path . $folder;
+			$it = new RecursiveDirectoryIterator($dir);
+			$files = new RecursiveIteratorIterator($it, RecursiveIteratorIterator::CHILD_FIRST);
+			foreach ($files as $file) {
+				if ($file->getFilename() === '.' || $file->getFilename() === '..') {
+					continue;
+				}
+
+				if ($file->isDir()) {
+					$this->rmdir($file->getRealPath());
+				} else {
+					$this->unlink($file->getRealPath());
+				}
+			}
+			$this->rmdir($dir);
+		}*/
 
 		return true;
 	}
@@ -448,8 +471,10 @@ class UploadBehavior extends ModelBehavior {
 		foreach ($this->settings[$model->alias] as $options) {
 			if ($options['deleteFolderOnDelete'] == true) {
 				$this->deleteFolder($model, $options['path']);
+				//return true;
 			}
 		}
+
 		return $result;
 	}
 
@@ -1011,10 +1036,10 @@ class UploadBehavior extends ModelBehavior {
 		}
 
 		if ($usePrefixStyle) {
-			return '{size}_{filename}';
+			return '{size}{filename}';
 		}
 
-		return '{filename}_{size}';
+		return '{filename}{size}';
 	}
 
 /**
